Very very hacky but:
```
$ /home/runner/work/matechs-effect/matechs-effect/node_modules/.bin/ts-node bench/index.ts
native x 29,133 ops/sec ±1.34% (82 runs sampled)
effect x 16,982 ops/sec ±1.65% (80 runs sampled)
wave x 13,941 ops/sec ±1.25% (82 runs sampled)
qio x 18,585 ops/sec ±2.08% (80 runs sampled)
Fastest is native
$ /home/runner/work/matechs-effect/matechs-effect/node_modules/.bin/ts-node bench/nestedMap.ts
effect x 3,604 ops/sec ±1.34% (79 runs sampled)
wave x 2,872 ops/sec ±1.20% (80 runs sampled)
qio x 6,388 ops/sec ±1.56% (80 runs sampled)
Fastest is qio
$ /home/runner/work/matechs-effect/matechs-effect/node_modules/.bin/ts-node bench/nestedChain.ts
effect x 625 ops/sec ±1.20% (82 runs sampled)
wave x 418 ops/sec ±1.37% (80 runs sampled)
qio x 680 ops/sec ±1.42% (81 runs sampled)
Fastest is qio
```
Basically in this way env adds no call overhead at all